feat: Update the project to Swift 3.0
The app now runs on devices using iOS 10 and can be worked
on using Xcode 8.0.  To run on a device, the info.plist file
needed to be edited to allow the app permission to access the
device's photo album and camera.  Storyboard was updated
so that item frames accurately reflect the appearance of
the app at runtime.